refactor(typescript): move some common components to TS
